# docs/002.plan/002.dev/Phase1/006.apix_l1_openapi_3_1.yaml
openapi: 3.1.0
info:
  title: APIX L1 Event & Risk Control API
  version: 1.0.0
  description: |
    APIX Avalanche L1(Chain ID 402) 이벤트 수집 및 AML/남용 탐지 운영 API.
    Proof-of-Truth는 온체인 이벤트이며, 본 API는 조회 및 운영 제어 기능을 제공합니다.
servers:
  - url: https://api.apix.example.com
    description: Production
  - url: https://staging-api.apix.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local
paths:
  /v1/health:
    get:
      summary: 시스템 상태 조회
      operationId: getHealth
      tags: [Admin]
      responses:
        '200':
          description: Health check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /v1/events:
    get:
      summary: 정규화된 이벤트 조회
      operationId: listEvents
      tags: [Events]
      parameters:
        - name: wallet
          in: query
          required: false
          schema:
            type: string
        - name: event_type
          in: query
          required: false
          schema:
            type: string
        - name: api_id
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: UTC ISO8601 timestamp
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: UTC ISO8601 timestamp
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, CONFIRMED, REVERTED, FAILED]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/risk/alerts:
    get:
      summary: Risk Alert 목록 조회
      operationId: listRiskAlerts
      tags: [Risk]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: grade
          in: query
          required: false
          schema:
            type: string
            enum: [NORMAL, WARN, SUSPECT, CRITICAL]
        - name: wallet
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, ACKNOWLEDGED, INVESTIGATING, RESOLVED]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Risk alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAlertListResponse'
  /v1/risk/score/{wallet}:
    get:
      summary: 지갑별 현재 위험 점수 조회
      operationId: getWalletRiskScore
      tags: [Risk]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: wallet
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Risk score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScoreResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/risk/alerts/{alert_id}/ack:
    post:
      summary: Risk Alert 승인/상태 업데이트
      operationId: acknowledgeRiskAlert
      tags: [Risk]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskAlertAckRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAlertResponse'
  /v1/ops/lag:
    get:
      summary: 수집/처리 lag 조회
      operationId: getOperationLag
      tags: [Ops]
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Lag snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LagSnapshotResponse'
  /v1/ops/replay:
    post:
      summary: 블록 구간 재수집/재처리 트리거
      operationId: triggerReplay
      tags: [Ops]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplayRequest'
      responses:
        '202':
          description: Replay accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplayResponse'
  /v1/ops/rules/reload:
    post:
      summary: 리스크 룰 재적재
      operationId: reloadRiskRules
      tags: [Ops]
      security:
        - ApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleReloadRequest'
      responses:
        '200':
          description: Reload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleReloadResponse'
  /webhooks/apix-l1:
    post:
      summary: AvaCloud Webhook 수신 (보조 채널)
      operationId: receiveWebhookEvent
      tags: [Webhook]
      parameters:
        - name: X-Webhook-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEnvelope'
      responses:
        '202':
          description: Accepted for reconciliation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    HealthResponse:
      type: object
      required: [status, checked_at]
      properties:
        status: { type: string, enum: [UP, DEGRADED, DOWN] }
        checked_at: { type: string, format: date-time }
        chain_id: { type: integer, example: 402 }
        websocket_connected: { type: boolean }
        kafka_healthy: { type: boolean }
        lag_ms: { type: number, format: float }
    L1Event:
      type: object
      required: [dedupe_key, chain_id, tx_hash, block_number, log_index, event_type]
      properties:
        event_id: { type: string }
        dedupe_key: { type: string, example: '402|0xabc...|14' }
        chain_id: { type: integer, example: 402 }
        block_number: { type: integer }
        block_hash: { type: string }
        tx_hash: { type: string }
        log_index: { type: integer }
        contract_address: { type: string }
        event_type: { type: string }
        wallet: { type: string }
        api_id: { type: string }
        plan_id: { type: string }
        amount: { type: string, description: Raw token amount as decimal string }
        token_symbol: { type: string, example: APIX }
        status: { type: string, enum: [PENDING, CONFIRMED, REVERTED, FAILED] }
        timestamp: { type: string, format: date-time }
        source: { type: string, enum: [wss, webhook] }
        risk_context: { type: object, additionalProperties: true }
    EventListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/L1Event' }
        page:
          $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      required: [page, size, total, has_more]
      properties:
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
        prev_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
    RiskAlert:
      type: object
      required: [alert_id, wallet, risk_score, risk_grade, status]
      properties:
        alert_id: { type: string }
        wallet: { type: string }
        risk_score: { type: number, minimum: 0, maximum: 100 }
        risk_grade: { type: string, enum: [NORMAL, WARN, SUSPECT, CRITICAL] }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, INVESTIGATING, RESOLVED] }
        rule_ids: { type: array, items: { type: string } }
        reasons: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        assigned_to: { type: string, nullable: true }
        ttl_sec: { type: integer }
    RiskAlertListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RiskAlert' }
        page:
          $ref: '#/components/schemas/PageInfo'
    RiskScoreResponse:
      type: object
      required: [wallet, risk_score, risk_grade]
      properties:
        wallet: { type: string }
        risk_score: { type: number, minimum: 0, maximum: 100 }
        risk_grade: { type: string, enum: [NORMAL, WARN, SUSPECT, CRITICAL] }
        signals: { type: object, additionalProperties: true }
        computed_at: { type: string, format: date-time }
        ttl_sec: { type: integer }
    RiskAlertResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/RiskAlert'
        - type: object
          properties:
            comment: { type: string }
    RiskAlertAckRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACKNOWLEDGED, INVESTIGATING, RESOLVED]
        note:
          type: string
          maxLength: 500
        assigned_to:
          type: string
    LagSnapshotResponse:
      type: object
      required: [kafka_lag, ws_lag_seconds, mismatch_count]
      properties:
        kafka_lag: { type: integer }
        ws_lag_seconds: { type: number, format: float }
        mismatch_count: { type: integer }
        latest_block: { type: integer }
        last_reconciled_block: { type: integer }
    ReplayRequest:
      type: object
      required: [from_block, to_block]
      properties:
        from_block: { type: integer, minimum: 0 }
        to_block: { type: integer, minimum: 0 }
        reason: { type: string, maxLength: 500 }
    ReplayResponse:
      type: object
      required: [job_id, accepted]
      properties:
        job_id: { type: string }
        accepted: { type: boolean }
        expected_records: { type: integer }
    RuleReloadRequest:
      type: object
      properties:
        dry_run:
          type: boolean
          default: false
        force:
          type: boolean
          default: false
    RuleReloadResponse:
      type: object
      required: [applied_version, rules_loaded]
      properties:
        applied_version: { type: string }
        rules_loaded: { type: integer }
        changed: { type: boolean }
        started_at: { type: string, format: date-time }
    WebhookEnvelope:
      type: object
      required: [event_id, payload, received_at]
      properties:
        source: { type: string, example: avacloud }
        event_type: { type: string }
        event_id: { type: string }
        payload:
          type: object
          additionalProperties: true
        received_at: { type: string, format: date-time }
        signature: { type: string }
    WebhookResponse:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }